test: relocate unit test
